Realizes storage abstraction for points, offers and destinations
