+++ src/api/destinations-storage.js
@@ -1,0 +1,9 @@
+import AbstractStorage from '../utils/abstract-storage';
+
+class DestinationsStorage extends AbstractStorage {
+  constructor(key, storage) {
+    super(key, storage);
+  }
+}
+
+export default DestinationsStorage;
